+++ _examples/asset/main.go
@@ -53,7 +53,7 @@ func main() {
 	// возьмем номер счета из .env
 	accountId, _ := os.LookupEnv("FINAM_ACCOUNT")
 	_ = accountId
-	symbol := "IMOEXF@RTSX" //"SBER@MISX" //"FEES@MISX" //"SBER@MISX" // "SIU5@RTSX" // "SBER@MISX" // "RU000A106L18@MISX"
+	symbol := "SBER@MISX" //"IMOEXF@RTSX" //"SBER@MISX" //"FEES@MISX" //"SBER@MISX" // "SIU5@RTSX" // "SBER@MISX" // "RU000A106L18@MISX"
 	_ = symbol
 	// Получение информации по конкретному инструменту
 	getAsset(ctx, assetService, accountId, symbol)