+++ asset.go
@@ -29,8 +29,9 @@ package finam
 
 import (
 	"context"
-	assets_service "github.com/Ruvad39/go-finam-grpc/proto/grpc/tradeapi/v1/assets"
 	"time"
+
+	assets_service "github.com/Ruvad39/go-finam-grpc/proto/grpc/tradeapi/v1/assets"
 )
 
 // AssetServiceClient клиент для работы AssetsService
@@ -47,6 +48,8 @@ func NewAssetServiceClient(c *Client) *AssetServiceClient {
 
 // GetTime вернем текущее время сервера (в TzMoscow)
 func (s *AssetServiceClient) GetTime(ctx context.Context) (time.Time, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	resp, err := s.AssetService.Clock(ctx, &assets_service.ClockRequest{})
 	if err != nil {
 		return time.Time{}, err
@@ -56,25 +59,35 @@ func (s *AssetServiceClient) GetTime(ctx context.Context) (time.Time, error) {
 
 // GetExchanges Получение списка доступных бирж, названия и mic коды
 func (s *AssetServiceClient) GetExchanges(ctx context.Context) (*assets_service.ExchangesResponse, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	return s.AssetService.Exchanges(ctx, &assets_service.ExchangesRequest{})
 }
 
 // GetAssets Получение списка доступных инструментов, их описание
 func (s *AssetServiceClient) GetAssets(ctx context.Context) (*assets_service.AssetsResponse, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	return s.AssetService.Assets(ctx, &assets_service.AssetsRequest{})
 }
 
 // GetAsset Получение информации по конкретному инструменту
 func (s *AssetServiceClient) GetAsset(ctx context.Context, accountId, symbol string) (*assets_service.GetAssetResponse, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	return s.AssetService.GetAsset(ctx, &assets_service.GetAssetRequest{AccountId: accountId, Symbol: symbol})
 }
 
 // GetAssetParams Получение торговых параметров по инструменту
 func (s *AssetServiceClient) GetAssetParams(ctx context.Context, accountId, symbol string) (*assets_service.GetAssetParamsResponse, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	return s.AssetService.GetAssetParams(ctx, &assets_service.GetAssetParamsRequest{AccountId: accountId, Symbol: symbol})
 }
 
 // GetSchedule Получение расписания торгов для инструмента
 func (s *AssetServiceClient) GetSchedule(ctx context.Context, symbol string) (*assets_service.ScheduleResponse, error) {
+	ctx, cancel := context.WithTimeout(ctx, s.client.opts.callTimeout)
+	defer cancel()
 	return s.AssetService.Schedule(ctx, &assets_service.ScheduleRequest{Symbol: symbol})
 }